requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==6.0
djangorestframework==3.16.1
drf-yasg==1.21.11
idna==3.11
inflection==0.5.1
packaging==25.0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
sqlparse==0.5.5
stripe==14.1.0
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.2
whitenoise==6.11.0
gunicorn